crud type,subcategory,measure and brand

+++ resources/views/admin/measure/index.blade.php
@@ -6,7 +6,7 @@
   
       <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
         <h1 class="h2">          
-          {{trans('app.categories')}}
+          {{trans('app.measures')}}
         </h1>
       </div>
   <br />    
@@ -22,8 +22,8 @@
     <tr>
       <th scope="col">#ID</th>
       <th scope="col">{{trans('app.name')}}</th>
+      <th scope="col">{{trans('validation.attributes.abbrev')}}</th>
       <th scope="col">Slug</th>
-      <th scope="col">{{trans('app.type')}}</th>
       <th scope="col">
         <span data-feather="settings"></span>
       </th>
@@ -34,12 +34,14 @@
     <tr>
       <th scope="row">{{$item->id}}</th>
       <td>{{$item->name}}</td>
+      <td>{{$item->abbrev}}</td>
       <td>{{$item->slug}}</td>
-      <td>{{$item->type->name}}</td>
       <td>
-        <a href="{{url('admin/category/'.$item->id.'/edit')}}" class="text-primary"><span data-feather="edit"></span></a>
-        <a href="#" class="text-danger" id="delete_item" data-id="{{ $item->id }}"><span data-feather="trash"></span></a>
-               
+        <a href="{{url('admin/measure/'.$item->id.'/edit')}}" class="text-primary"><span data-feather="edit"></span></a>
+        
+        @if(!$item->products()->exists())
+          <a href="#" class="text-danger" id="delete_item" data-id="{{ $item->id }}"><span data-feather="trash"></span></a>
+        @endif       
       </td>
     </tr>
     @endforeach
@@ -62,16 +64,16 @@
   jQuery(document).ready(function(){
 
     $("a#delete_item").click(function(e){
-        
+              
         var id = $(this).data("id");
         e.preventDefault();
         var token = $("meta[name='csrf-token']").attr("content");
         var url = e.target;
-        
+                
         if (confirm("{{trans('app.are_you_sure')}}")) {
           
           $.ajax({
-            url: "category/"+id, //or you can use url: "company/"+id,
+            url: "measure/"+id, //or you can use url: "company/"+id,
             type: 'DELETE',
             dataType: "json",
             data: {
